+++ R/Framework.R
@@ -697,7 +697,6 @@ saveAsProject <- function(
 #' 
 copyProject <- function(projectPath, newProjectPath, ow = FALSE, empty.output = FALSE, empty.input = FALSE, empty.memory = FALSE, empty.processData = FALSE, processDataToBeEmptied = NULL, close = FALSE, save = NULL, msg = TRUE) {
     
-    
     # Check whether the project to be used as template exists:
     if(!dir.exists(projectPath) || !isProject(projectPath)) {
         stop("The path ", projectPath, " does not point to a StoX project.")
@@ -713,9 +712,10 @@ copyProject <- function(projectPath, newProjectPath, ow = FALSE, empty.output = 
         }
     }
     
-    if(close) {
+    if(close && isOpenProject(projectPath)) {
         closeProject(projectPath, save = save, msg = msg)
     }
+    
     
     #suppressWarnings(dir.create(newProjectPath, recursive = TRUE))
     createProjectSkeleton(newProjectPath, ow = ow)
@@ -830,6 +830,11 @@ copyProject <- function(projectPath, newProjectPath, ow = FALSE, empty.output = 
     # Make sure the project session is complete:
     if(isOpenProject(newProjectPath)) {
         createProjectSessionFolderStructure(newProjectPath)
+    }
+    
+    # Close the new project if the old is closed, or has been closed earlier in the function:
+    if(!isOpenProject(projectPath)) {
+        closeProject(newProjectPath, save = save, msg = msg)
     }
     
     
@@ -1076,7 +1081,10 @@ readProjectDescription <- function(
                 projectDescriptionFile = tempProjectDescriptionFile, 
                 optionalDependencies = TRUE
             )
+            
             valid <- validateProjectDescriptionFile(tempProjectDescriptionFile)
+            # Delete the temporary project description file:
+            unlink(tempProjectDescriptionFile)
         #}
         
         if(!isTRUE(valid)) {
@@ -4471,7 +4479,8 @@ formatFunctionParameters <-  function(functionParameters, functionName, projectP
             
             # Warning if there are parameters not specified in the function definition:
             if(length(invalid)) {
-                warning("StoX: The following functionParameters are not specified in the definition of function ", functionName, ": ", paste(invalid, collapse = ", "))
+                warning("StoX: The following functionParameters are not specified in the definition of function ", functionName, ", and were removed: ", paste(invalid, collapse = ", "))
+                functionParameters <- functionParameters[present]
             }
             # Change class to the defined class:
             else if(length(present)) {
@@ -4910,7 +4919,7 @@ expandProcess <- function(projectPath, modelName, processName, values = NULL, re
         )
     }
     
-    # Add an AddToStoxBiotic process which inputs the firstStoxBioticProcess 
+    # Add the process:
     suppressWarnings(
         addProcess(
             projectPath = projectPath, 
@@ -5202,6 +5211,16 @@ runProcess <- function(
     }
     if(is.list(replaceData) && !data.table::is.data.table(replaceData) && is.character(replaceData$FunctionName)) {
         
+        # Add a message when replacing data using a function:
+        message(
+            "StoX: Replacing data in ", modelName, " process ", 
+            getProcessIndexFromProcessID(projectPath = projectPath, modelName = modelName, processID = processID), 
+            ": ", 
+            getProcessName(projectPath = projectPath, modelName = modelName, processID = processID), 
+            " using the function ", replaceData$FunctionName, ".",
+            appendLF = TRUE
+        )
+        
         #if(!exists(replaceData$FunctionName)) {
         #    stop("If replaceData is given as a list with a function name first, this must be an existing function (was ", replaceData$FunctionName, ").")
         #}
@@ -5231,6 +5250,17 @@ runProcess <- function(
         )
     }
     else if(length(replaceData)) {
+        # Add a message when replacing data:
+        message(
+            "StoX: Replacing data in ", modelName, " process ", 
+            getProcessIndexFromProcessID(projectPath = projectPath, modelName = modelName, processID = processID), 
+            ": ", 
+            getProcessName(projectPath = projectPath, modelName = modelName, processID = processID), 
+            " using the replaceData option.", 
+            appendLF = TRUE
+        )
+        
+        # Do the actual replacement:
         processOutput <- replaceData
     }
     
@@ -5394,6 +5424,17 @@ getFunctionArguments <- function(projectPath, modelName, processID, arguments = 
         warning("The replaceArgs/replaceArgsList contains the following parameters that do not exist in the process: ", paste(namesOfReplaceArgsNotToInsert, sep = ", "))
     }
     if(length(namesOfReplaceArgsToInsert)) {
+        # Add a message when replacing data:
+        message(
+            "StoX: Replacing the following arguments in ", modelName, " process ", 
+            getProcessIndexFromProcessID(projectPath = projectPath, modelName = modelName, processID = processID), 
+            ": ", 
+            getProcessName(projectPath = projectPath, modelName = modelName, processID = processID), 
+            ": ", 
+            paste(namesOfReplaceArgsToInsert, collapse = ", "), ".",
+            appendLF = TRUE
+        )
+        
         functionArguments[namesOfReplaceArgsToInsert] <- replaceArgs[namesOfReplaceArgsToInsert]
     }
     
@@ -6416,7 +6457,12 @@ writeProcessOutputTextFile <- function(processOutput, projectPath, modelName, pr
                 file.ext = outputFileType
             )
             
+            
             file.copy(processOutput[[1]], filePath)
+            
+            # NO THIS IS NOT SMART. WE CANNOT DELETE THE MEMORY FILE:
+            # delete the temporary bootstrap files:
+            #unlink(processOutput[[1]])
             
             message(
                 "StoX: The bootstrap file can be read into R using the following command:", "\n",